+++ packages/v2/agent/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @copilotkitnext/agent
 
+## 1.53.0
+
+### Patch Changes
+
+- 1510f64: feat: enable mcp and a2ui middleware directly from copilotkit runtime
+
+## 1.53.0-next.6
+
+## 1.53.0-next.5
+
+## 1.52.2-next.4
+
+## 1.52.2-next.3
+
+### Patch Changes
+
+- 1510f64: feat: enable mcp and a2ui middleware directly from copilotkit runtime
+
 ## 1.52.2-next.2
 
 ## 1.52.2-next.1